+++ tests/test_nosql_dataset_utils.py
@@ -26,6 +26,8 @@ def test_get_readme_from_uri(tmp_app):  # NOQA
         "type": "dataset",
         "readme": {"description": "test dataset"},
         "creator_username": "olssont",
+        "frozen_at": 1536238185.881941,
+        "created_at": 1536236399.19497,
     }
 
     with pytest.raises(ValidationError):