+++ tests/test_sql_list_datasets_by_user.py
@@ -33,6 +33,8 @@ def test_list_datasets_by_user(tmp_app):  # NOQA
         "uri": uri_1,
         "name": "ds_1",
         "creator_username": "olssont",
+        "frozen_at": 1536238185.881941,
+        "created_at": 1536236399.19497,
     }
     admin_metadata_2 = {
         "base_uri": base_uri_2,
@@ -40,6 +42,8 @@ def test_list_datasets_by_user(tmp_app):  # NOQA
         "uri": uri_2,
         "name": "ds_2",
         "creator_username": "olssont",
+        "frozen_at": 1536238185.881941,
+        "created_at": 1536236399.19497,
     }
     register_dataset_admin_metadata(admin_metadata_1)
     register_dataset_admin_metadata(admin_metadata_2)