+++ tests/test_utils_dataset_info_is_valid.py
@@ -13,6 +13,7 @@ def test_dataset_info_is_valid_returns_true_on_valid_info():
         "readme": {"description": "test dataset"},
         "base_uri": "file:///tmp",
         "creator_username": "olssont",
+        "frozen_at": 1536238185.881941,
     }
     assert dataset_info_is_valid(info)
 
@@ -27,6 +28,7 @@ def test_dataset_info_returns_false_when_key_data_is_missing():
         "readme": {"description": "test dataset"},
         "base_uri": "file:///tmp",
         "creator_username": "olssont",
+        "frozen_at": 1536238185.881941,
     }
     assert not dataset_info_is_valid(info)
 
@@ -37,6 +39,7 @@ def test_dataset_info_returns_false_when_key_data_is_missing():
         "readme": {"description": "test dataset"},
         "base_uri": "file:///tmp",
         "creator_username": "olssont",
+        "frozen_at": 1536238185.881941,
     }
     assert not dataset_info_is_valid(info)
 
@@ -47,6 +50,7 @@ def test_dataset_info_returns_false_when_key_data_is_missing():
         "readme": {"description": "test dataset"},
         "base_uri": "file:///tmp",
         "creator_username": "olssont",
+        "frozen_at": 1536238185.881941,
     }
     assert not dataset_info_is_valid(info)
 
@@ -57,6 +61,7 @@ def test_dataset_info_returns_false_when_key_data_is_missing():
         "readme": {"description": "test dataset"},
         "base_uri": "file:///tmp",
         "creator_username": "olssont",
+        "frozen_at": 1536238185.881941,
     }
     assert not dataset_info_is_valid(info)
 
@@ -67,6 +72,7 @@ def test_dataset_info_returns_false_when_key_data_is_missing():
         "readme": {"description": "test dataset"},
         "name": "my-dataset",
         "creator_username": "olssont",
+        "frozen_at": 1536238185.881941,
     }
     assert not dataset_info_is_valid(info)
 
@@ -76,6 +82,7 @@ def test_dataset_info_returns_false_when_key_data_is_missing():
         "name": "my-dataset",
         "base_uri": "file:///tmp",
         "creator_username": "olssont",
+        "frozen_at": 1536238185.881941,
     }
     assert not dataset_info_is_valid(info)
 
@@ -87,6 +94,19 @@ def test_dataset_info_returns_false_when_key_data_is_missing():
         "name": "my-dataset",
         "readme": {"description": "test dataset"},
         "base_uri": "file:///tmp",
+        "frozen_at": 1536238185.881941,
+    }
+    assert not dataset_info_is_valid(info)
+
+    # No frozen_at.
+    info = {
+        "uuid": "af6727bf-29c7-43dd-b42f-a5d7ede28337",
+        "type": "dataset",
+        "uri": "file:///tmp/a_dataset",
+        "name": "my-dataset",
+        "readme": {"description": "test dataset"},
+        "base_uri": "file:///tmp",
+        "creator_username": "olssont",
     }
     assert not dataset_info_is_valid(info)
 
@@ -98,7 +118,11 @@ def test_dataset_info_returns_false_when_type_is_not_dataset():
         "uuid": "af6727bf-29c7-43dd-b42f-a5d7ede28337",
         "type": "protodataset",
         "uri": "file:///tmp/a_dataset",
+        "name": "my-dataset",
+        "readme": {"description": "test dataset"},
+        "base_uri": "file:///tmp",
         "creator_username": "olssont",
+        "frozen_at": 1536238185.881941,
     }
     assert not dataset_info_is_valid(info)
 
@@ -110,7 +134,11 @@ def test_dataset_info_returns_false_if_uuid_looks_invalid():
         "uuid": "af6727bf-29c7-43dd-b42f",
         "type": "protodataset",
         "uri": "file:///tmp/a_dataset",
+        "name": "my-dataset",
+        "readme": {"description": "test dataset"},
+        "base_uri": "file:///tmp",
         "creator_username": "olssont",
+        "frozen_at": 1536238185.881941,
     }
     assert not dataset_info_is_valid(info)
 
@@ -118,13 +146,14 @@ def test_dataset_info_returns_false_if_uuid_looks_invalid():
 def test_dataset_info_is_valid_returns_false_if_base_uri_ends_with_slash():
     from dtool_lookup_server.utils import dataset_info_is_valid
 
-    # Below is the minimum data required to register a dataset.
     info = {
         "uuid": "af6727bf-29c7-43dd-b42f-a5d7ede28337",
         "type": "dataset",
         "uri": "file:///tmp/a_dataset",
-        "name": "my-dataset",
         "base_uri": "file:///tmp/",
         "creator_username": "olssont",
+        "name": "my-dataset",
+        "readme": {"description": "test dataset"},
+        "frozen_at": 1536238185.881941,
     }
     assert not dataset_info_is_valid(info)