+++ tests/test_utils_register_dataset.py
@@ -41,6 +41,8 @@ def test_register_dataset(tmp_app):   # NOQA
         "type": "dataset",
         "readme": {"description": "test dataset"},
         "creator_username": "olssont",
+        "frozen_at": 1536238185.881941,
+        "created_at": 1536236399.19497,
     }
 
     register_dataset(dataset_info)
@@ -51,6 +53,66 @@ def test_register_dataset(tmp_app):   # NOQA
         "uri": uri,
         "name": "my-dataset",
         "creator_username": "olssont",
+        "frozen_at": 1536238185.881941,
+        "created_at": 1536236399.19497,
+    }
+    assert get_admin_metadata_from_uri(uri) == expected_content
+    assert get_readme_from_uri(uri) == dataset_info["readme"]
+
+    with pytest.raises(ValidationError):
+        register_dataset({"name": "not-all-required-metadata"})
+
+
+def test_register_dataset_without_created_at(tmp_app):   # NOQA
+    from dtool_lookup_server import ValidationError
+    from dtool_lookup_server.utils import (
+        register_users,
+        register_base_uri,
+        update_permissions,
+        register_dataset,
+        get_admin_metadata_from_uri,
+        get_readme_from_uri,
+    )
+
+    register_users([
+        dict(username="grumpy"),
+        dict(username="sleepy"),
+    ])
+
+    base_uri = "s3://snow-white"
+    register_base_uri(base_uri)
+
+    permissions = {
+        "base_uri": base_uri,
+        "users_with_search_permissions": ["grumpy", "sleepy"],
+        "users_with_register_permissions": ["grumpy"],
+    }
+    update_permissions(permissions)
+
+    uuid = "af6727bf-29c7-43dd-b42f-a5d7ede28337"
+    uri = "{}/{}".format(base_uri, uuid)
+    dataset_info = {
+        "base_uri": base_uri,
+        "uuid": uuid,
+        "uri": uri,
+        "name": "my-dataset",
+        "type": "dataset",
+        "readme": {"description": "test dataset"},
+        "creator_username": "olssont",
+        "frozen_at": 1536238185.881941,
+    }
+
+    register_dataset(dataset_info)
+
+    # When missing, created_at will be set to frozen_at.
+    expected_content = {
+        "base_uri": base_uri,
+        "uuid": uuid,
+        "uri": uri,
+        "name": "my-dataset",
+        "creator_username": "olssont",
+        "frozen_at": 1536238185.881941,
+        "created_at": 1536238185.881941,
     }
     assert get_admin_metadata_from_uri(uri) == expected_content
     assert get_readme_from_uri(uri) == dataset_info["readme"]